+++ esibd/devices/keithley/keithley.py
@@ -61,7 +61,7 @@ class CurrentChannel(Channel):
     ADDRESS    = 'Address'
     VOLTAGE    = 'Voltage'
 
-    def getDefaultChannel(self) -> None:
+    def getDefaultChannel(self) -> dict[str, dict]:
         channel = super().getDefaultChannel()
         channel[self.VALUE][Parameter.HEADER] = 'I (pA)'
         channel[self.CHARGE] = parameterDict(value=0, parameterType=PARAMETERTYPE.FLOAT, advanced=False, header='C (pAh)', indicator=True, attr='charge')