+++ esibd/devices/ni9263/ni9263.py
@@ -34,7 +34,7 @@ class VoltageChannel(Channel):
 
     ADDRESS = 'Address'
 
-    def getDefaultChannel(self) -> None:
+    def getDefaultChannel(self) -> dict[str, dict]:
         channel = super().getDefaultChannel()
         channel[self.VALUE][Parameter.HEADER] = 'Voltage (V)'  # overwrite to change header
         channel[self.MIN][Parameter.VALUE] = 0