+++ esibd/devices/rspd3303c/rspd3303c.py
@@ -88,7 +88,7 @@ class VoltageChannel(Channel):
     POWER     = 'Power'
     ID        = 'ID'
 
-    def getDefaultChannel(self) -> None:
+    def getDefaultChannel(self) -> dict[str, dict]:
         channel = super().getDefaultChannel()
         channel[self.VALUE][Parameter.HEADER] = 'Voltage (V)'  # overwrite to change header
         channel[self.MIN][Parameter.VALUE] = 0